ISSUE #1339 - Renamed files for clarity changed apikey middleware to actual middleware, merge apikey and profile route files into one file
